Name change mwoperators -> operators
