+++ api/MWOperators
@@ -1,8 +1,8 @@
-#include "mwoperators/PoissonOperator.h"
-#include "mwoperators/HelmholtzOperator.h"
-#include "mwoperators/DerivativeConvolution.h"
-#include "mwoperators/IdentityConvolution.h"
-#include "mwoperators/ABGVOperator.h"
-#include "mwoperators/PHOperator.h"
+#include "operators/PoissonOperator.h"
+#include "operators/HelmholtzOperator.h"
+#include "operators/DerivativeConvolution.h"
+#include "operators/IdentityConvolution.h"
+#include "operators/ABGVOperator.h"
+#include "operators/PHOperator.h"
 
 #include "treebuilders/apply.h"
