+++ src/operators/MWOperator.h
