+++ src/operators/PoissonKernel.cpp
@@ -10,41 +10,48 @@
 
 #include <cmath>
 
-#include "mwoperators/HelmholtzKernel.h"
+#include "operators/PoissonKernel.h"
 #include "functions/GaussFunc.h"
 #include "mwutils/Printer.h"
 
-using namespace std;
-
 namespace mrcpp {
 
-/** generate an approximation of the 3d helmholtz kernel expanded in gaussian functions
+/** generate an approximation of the 3d poisson kernel expanded in
+ * gaussian functions this routine assumes that the expansion be centered
  */
-void HelmholtzKernel::initializeKernel() {
+void PoissonKernel::initializeKernel() {
     //Constructed on [rMin/rMax, 1.0], and then rescaled to [rMin,rMax]
     double r0 = this->rMin/this->rMax;
     double r1 = this->rMax;
-    double mu_tilde = this->mu*r1;
+
+    double t1 = 1.0L;
+    while ((2 * t1 * exp(-t1)) > this->epsilon) {
+        t1 *= 1.1L;
+    }
+    double t2 = 1.0L;
+    while ((sqrt(t2) * exp(-t2) / r0) > this->epsilon) {
+        t2 *= 1.1L;
+    }
 
     // Set the truncation limits s1,s2 of the integral (integrate over [s1,s2])
     // for achieving relative error epsilon
-    double t = max((-2.5L * log(this->epsilon)), 5.0L);
-    double s1 = -log(4 * t / (mu_tilde * mu_tilde)) / 2;
-    double s2 = log(t / (r0 * r0)) / 2;
+    double s1 = -log(2 * t1);
+    double s2 = log(t2 / (r0 * r0)) / 2;
 
-    // Now, set the proper step size h for use in the trapezoidal rule
-    // for given MU
-    double h = 1.0 / (0.20L - 0.47L * log10(this->epsilon));
+    // Now, set the step size h for use in the trapezoidal rule for given MU
+    double h = 1 / (0.2L - 0.47L * log10(this->epsilon));
     int n_exp = (int) ceil((s2 - s1) / h) + 1;
+
     if (n_exp > MaxSepRank) MSG_FATAL("Maximum separation rank exceeded.");
 
     for (int i = 0; i < n_exp; i++) {
         double arg = s1 + h * i;
-        double temp = -arg * 2.0;
-        double temp2 = -mu_tilde * mu_tilde * exp(temp) / 4.0 + arg;
-        double beta = (h * (2.0 / root_pi) * exp(temp2));
-        double temp3 = 2.0L * arg;
-        double alpha = exp(temp3);
+        double sinharg = sinh(arg);
+        double cosharg = cosh(arg);
+        double onepexp = 1.0 + exp(-sinharg);
+
+        double alpha = 4.0L * (sinharg+log(onepexp)) * (sinharg+log(onepexp));
+        double beta = h * (4.0L / root_pi) * cosharg / onepexp;
 
         alpha *= 1.0/(r1*r1);
         beta *= 1.0/r1;
@@ -58,12 +65,11 @@ void HelmholtzKernel::initializeKernel() {
     this->calcSquareNorm();
 }
 
-std::ostream& HelmholtzKernel::print(std::ostream &o) const {
-    o << " HelmholtzKernel: " << std::endl;
+std::ostream& PoissonKernel::print(std::ostream &o) const {
+    o << " PoissonKernel: " << std::endl;
     o << " epsilon:  " << this->epsilon << std::endl;
     o << " rMin:     " << this->rMin << std::endl;
     o << " rMax:     " << this->rMax << std::endl;
-    o << " mu:       " << this->mu << std::endl;
     return o;
 }
