+++ src/treebuilders/DerivativeCalculator.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "treebuilders/TreeCalculator.h"
-#include "mwoperators/OperatorStatistics.h"
+#include "operators/OperatorStatistics.h"
 
 namespace mrcpp {
