+++ src/treebuilders/apply.cpp
@@ -5,8 +5,8 @@
 #include "treebuilders/DefaultCalculator.h"
 #include "treebuilders/DerivativeCalculator.h"
 #include "treebuilders/ConvolutionCalculator.h"
-#include "mwoperators/DerivativeOperator.h"
-#include "mwoperators/ConvolutionOperator.h"
+#include "operators/DerivativeOperator.h"
+#include "operators/ConvolutionOperator.h"
 #include "mwtrees/FunctionTree.h"
 #include "mwutils/Printer.h"
 #include "mwutils/Timer.h"
